+++ js/util.js
@@ -1,0 +1,72 @@
+const checkNumber = (a, b) => {
+  if (Math.sign(a) === -1 || Math.sign(b) === -1) {
+    throw new Error('Negative number is not allowed');
+  }
+};
+
+const getRandomPositiveInteger = (a, b) => {
+  try {
+    checkNumber(a, b);
+  } catch (err) {
+    // eslint-disable-next-line
+    console.log(err);
+  }
+  const min = Math.ceil(Math.min(Math.abs(a), Math.abs(b)));
+  const max = Math.floor(Math.max(Math.abs(a), Math.abs(b)));
+  const result = Math.random() * (max - min + 1) + min;
+  return Math.floor(result);
+};
+
+const getRandomPositiveFloat = (a, b, digits = 1) => {
+  try {
+    checkNumber(a, b);
+  } catch (err) {
+    // eslint-disable-next-line
+    console.log(err);
+  }
+  const min = Math.min(Math.abs(a), Math.abs(b));
+  const max = Math.max(Math.abs(a), Math.abs(b));
+  const result = Math.random() * (max - min) + min;
+  return +result.toFixed(digits);
+};
+
+const putZeroBeforeNumber = (number) => {
+  if (String(number).length === 1) {
+    return `0${number}`;
+  }
+  return number.toString();
+};
+
+const copyMixArrayRandomSize = (arr) => {
+  const arrayCopy = arr.slice();
+  const arrayRandomSize = getRandomPositiveInteger(1, arrayCopy.length);
+  const result = new Array();
+  for (let i = 0; i  < arrayRandomSize; i++) {
+    const random = getRandomPositiveInteger(0, arrayCopy.length - 1);
+    const elem = arrayCopy.splice(random, 1)[0];
+    result.push(elem);
+  }
+  return result;
+};
+
+const getArrayRandomInteger = (size) => {
+  const arr = new Array();
+  const result = new Array();
+  for (let i = size; i > 0; i--) {
+    arr.unshift(i);
+  }
+  while (arr.length > 0) {
+    const random = getRandomPositiveInteger(0, arr.length - 1);
+    const elem = arr.splice(random, 1)[0];
+    result.push(elem);
+  }
+  return result;
+};
+
+export {
+  getRandomPositiveInteger,
+  getRandomPositiveFloat,
+  putZeroBeforeNumber,
+  copyMixArrayRandomSize,
+  getArrayRandomInteger,
+};
